script: Move IndexedDB DOM interfaces to `script/dom/indexeddb/`
