+++ components/script/dom/indexeddb/idbdatabase.rs
@@ -27,9 +27,9 @@ use crate::dom::domstringlist::DOMStringList;
 use crate::dom::event::{Event, EventBubbles, EventCancelable};
 use crate::dom::eventtarget::EventTarget;
 use crate::dom::globalscope::GlobalScope;
-use crate::dom::idbobjectstore::IDBObjectStore;
-use crate::dom::idbtransaction::IDBTransaction;
-use crate::dom::idbversionchangeevent::IDBVersionChangeEvent;
+use crate::dom::indexeddb::idbobjectstore::IDBObjectStore;
+use crate::dom::indexeddb::idbtransaction::IDBTransaction;
+use crate::dom::indexeddb::idbversionchangeevent::IDBVersionChangeEvent;
 use crate::indexed_db::is_valid_key_path;
 use crate::script_runtime::CanGc;
