+++ components/script/dom/indexeddb/idbfactory.rs
@@ -12,7 +12,7 @@ use crate::dom::bindings::reflector::{DomGlobal, Reflector, reflect_dom_object};
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::str::DOMString;
 use crate::dom::globalscope::GlobalScope;
-use crate::dom::idbopendbrequest::IDBOpenDBRequest;
+use crate::dom::indexeddb::idbopendbrequest::IDBOpenDBRequest;
 use crate::indexed_db::convert_value_to_key;
 use crate::script_runtime::CanGc;
