+++ components/script/dom/indexeddb/idbrequest.rs
@@ -34,10 +34,10 @@ use crate::dom::domexception::DOMException;
 use crate::dom::event::{Event, EventBubbles, EventCancelable};
 use crate::dom::eventtarget::EventTarget;
 use crate::dom::globalscope::GlobalScope;
-use crate::dom::idbcursor::{IterationParam, iterate_cursor};
-use crate::dom::idbcursorwithvalue::IDBCursorWithValue;
-use crate::dom::idbobjectstore::IDBObjectStore;
-use crate::dom::idbtransaction::IDBTransaction;
+use crate::dom::indexeddb::idbcursor::{IterationParam, iterate_cursor};
+use crate::dom::indexeddb::idbcursorwithvalue::IDBCursorWithValue;
+use crate::dom::indexeddb::idbobjectstore::IDBObjectStore;
+use crate::dom::indexeddb::idbtransaction::IDBTransaction;
 use crate::indexed_db::key_type_to_jsval;
 use crate::realms::enter_realm;
 use crate::script_runtime::{CanGc, JSContext as SafeJSContext};
