+++ components/script/dom/mod.rs
@@ -333,17 +333,8 @@ pub(crate) mod headers;
 pub(crate) mod history;
 pub(crate) mod html;
 pub(crate) use self::html::*;
-pub(crate) mod idbcursor;
-pub(crate) mod idbcursorwithvalue;
-pub(crate) mod idbdatabase;
-pub(crate) mod idbfactory;
-pub(crate) mod idbindex;
-pub(crate) mod idbkeyrange;
-pub(crate) mod idbobjectstore;
-pub(crate) mod idbopendbrequest;
-pub(crate) mod idbrequest;
-pub(crate) mod idbtransaction;
-pub(crate) mod idbversionchangeevent;
+pub(crate) mod indexeddb;
+pub(crate) use self::indexeddb::*;
 pub(crate) mod inputevent;
 pub(crate) mod intersectionobserver;
 pub(crate) mod intersectionobserverentry;
